wz/hfu1/pyproject.toml
[project]
name = "hfu1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.31.4",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
